added a preview box for the content from editor
