+++ src/main.jsx
@@ -2,6 +2,7 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import App from './App.jsx'
 import './index.css'
+import Edit from './Components/edit.jsx'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
